add coverage stats using bedtools (experimental)
We are testing a new way to calculate coverage
across region (mainly for exome data) that
gives more information: different completeness values,
total coverage, mean and std ...

Beside that it will create a R template with a lot of
basic metrics already done in qcsummary step.

This is experimental and will run when `coverage_experimental`: true.
Not adding anything to docs yet.